WIP Abstract interfaces for 32/64 bit Mips cpus

+++ src/Archs/MIPS/Zarem.Debugger.MIPS/MipsDebugHandler.cs
@@ -83,7 +83,7 @@ public class MipsDebugHandler : IDebugHandler
     public ulong GetStepOutAddress(IComputer computer)
     {
         var mipsCpu = (MipsCpu)computer.Cpu;
-        return mipsCpu[GPRegister.ReturnAddress];
+        return mipsCpu[(int)GPRegister.ReturnAddress];
     }
 
     /// <inheritdoc/>
